genre-service/requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.10
starlette==0.26.1
typing==3.7.4.3
typing_extensions==4.5.0
uvicorn==0.21.1
